refactor: make Runner easy to start/stop for tests
