+++ modules/runner/src/main.ts
@@ -1,293 +1,13 @@
-import { execFile } from 'child_process';
 import debug from 'debug';
-import got from 'got';
-import { URL } from 'url';
-import { v4 as uuidv4 } from 'uuid';
-import { env } from '@electron/bugbot-shared/lib/env-vars';
-import {
-  FiddleBisectResult,
-  parseFiddleBisectOutput,
-} from './fiddle-bisect-parser';
+import { Runner } from './runner';
 
 const d = debug('runner');
 
-interface BaseJob {
-  client_data?: string;
-  gist: string;
-  id: string;
-  os?: 'darwin' | 'linux' | 'win32';
-  error?: string;
-  time_added: number;
-}
-
-interface Result {
-  bisect_range?: [string, string];
-  error?: string;
-  runner: string;
-  status: 'failure' | 'success' | 'system_error' | 'test_error';
-  time_begun: number;
-  time_ended: number;
-}
-
-interface BisectJob extends BaseJob {
-  type: 'bisect';
-  range: [string, string];
-}
-
-interface TestJob extends BaseJob {
-  type: 'test';
-  version: string;
-}
-
-type AnyJob = BisectJob | TestJob;
-
-type JsonPatch = unknown;
-
-/**
- * Returns a promise that resolves after the specified timeout.
- */
-function timeout(ms: number): Promise<void> {
-  return new Promise((resolve) => {
-    setTimeout(() => resolve(), ms);
-  });
-}
-
-class Runner {
-  private readonly uuid: string;
-  private readonly fiddleExecPath: string;
-  private readonly brokerUrl: string;
-  private readonly platform: string;
-  private readonly pollTimeoutMs: number;
-  private time_begun: number;
-
-  /**
-   * Creates and initializes the runner from environment variables and default
-   * values, then starts the runner's execution loop.
-   */
-  static start(): Promise<never> {
-    // Determine the OS filter from the current running platform
-    const { platform } = process;
-    if (!['darwin', 'linux', 'win32'].includes(platform)) {
-      d(`Unsupported platform '${platform}'; exiting.`);
-      return null;
-    }
-
-    // Create the runner
-    const runner: Runner = Object.create(Runner.prototype, {
-      brokerUrl: {
-        value: env('BUGBOT_BROKER_URL'),
-      },
-      fiddleExecPath: {
-        value: env('FIDDLE_EXEC_PATH'),
-      },
-      platform: {
-        value: platform,
-      },
-      pollTimeoutMs: {
-        value: 20 * 1000, // 20 seconds
-      },
-      uuid: {
-        value: uuidv4(),
-      },
-    });
-
-    // Begin running the poll loop
-    return runner.pollLoop();
-  }
-
-  async pollLoop(): Promise<never> {
-    // Wrap the loop in a try-catch to log errors and make the runner resilient
-    try {
-      // Check for any claimable jobs
-      const claimableJobs = await this.fetchUnclaimedJobs();
-
-      // If there are no unclaimed jobs then sleep and try again
-      if (claimableJobs.length === 0) {
-        await timeout(this.pollTimeoutMs);
-        return this.pollLoop();
-      }
-
-      // Otherwise, claim the first job available
-      const [jobId] = claimableJobs;
-      // TODO(clavin): would adding jitter (e.g. claim first OR second randomly)
-      // help reduce any possible contention?
-      let etag = '';
-      const [job, initalEtag] = await this.fetchJobAndEtag(jobId);
-      etag = initalEtag;
-
-      // Claim the job
-      this.time_begun = Date.now();
-      const current = {
-        runner: this.uuid,
-        time_begun: this.time_begun,
-      };
-      etag = await this.patchJobAndUpdateEtag(job.id, etag, [
-        {
-          op: 'replace',
-          path: '/current',
-          value: current,
-        },
-      ]);
-
-      // Another layer of catching errors to also unclaim the job if we error
-      try {
-        // Then determine how to run the job and return results
-        if (job.type === 'bisect') {
-          // Run the bisect
-          const res = await this.runBisect(job.range, job.gist);
-
-          // Report the result back to the job
-          const result: Result = {
-            runner: this.uuid,
-            status: 'success',
-            time_begun: this.time_begun,
-            time_ended: Date.now(),
-          };
-
-          if (res.success) {
-            result.bisect_range = [res.goodVersion, res.badVersion];
-          } else {
-            // TODO: distinguish between system_error (need maintainer attn)
-            // and test_error (implies user should revise test code).
-            // Examples:
-            // - child_process timeout: test_error
-            // - invalid gist id: test_error
-            // - failure to launch electron-fiddle: system error
-            result.status = 'system_error';
-            // TODO(clavin): ^ better wording
-            result.error = 'Failed to narrow test down to two versions';
-          }
-
-          etag = await this.patchJobAndUpdateEtag(job.id, etag, [
-            {
-              op: 'add',
-              path: '/history/-',
-              result,
-            },
-            {
-              op: 'replace',
-              path: '/last',
-              result,
-            },
-            {
-              op: 'remove',
-              path: '/current',
-            },
-          ]);
-
-          // } else if (job.type === 'test') {
-          // TODO
-        } else {
-          throw new Error(`unexpected job type: "${(job as AnyJob).type}"`);
-        }
-      } catch (err) {
-        // Unclaim the job and rethrow
-        // FIXME: should append to history here and set
-        // system_error or test_error based on err type
-        await this.patchJobAndUpdateEtag(job.id, etag, [
-          {
-            op: 'remove',
-            path: '/current',
-            value: Date.now(),
-          },
-        ]);
-        throw err;
-      }
-    } catch (err) {
-      d('error while polling broker: %O', err);
-    }
-
-    // Sleep and then try again
-    await timeout(this.pollTimeoutMs);
-    return this.pollLoop();
-  }
-
-  /**
-   * Polls the broker for a list of unclaimed job IDs.
-   */
-  private async fetchUnclaimedJobs(): Promise<string[]> {
-    // Craft the url to the broker
-    const jobs_url = new URL('api/jobs', this.brokerUrl);
-    // find jobs compatible with this runner...
-    jobs_url.searchParams.append('platform', `${this.platform}|undefined`);
-    // ...is not currently claimed
-    jobs_url.searchParams.append('current.runner', 'undefined');
-    // ...and which have never been run
-    jobs_url.searchParams.append('last.status', 'undefined');
-
-    // Make the request and return its response
-    return await got(jobs_url).json();
-  }
-
-  private async fetchJobAndEtag(id: string): Promise<[AnyJob, string]> {
-    const job_url = new URL(`api/jobs/${id}`, this.brokerUrl);
-    const resp = await got(job_url);
-
-    // Extract the etag header & make sure it was defined
-    const { etag } = resp.headers;
-    if (!etag) {
-      throw new Error('missing etag in broker job response');
-    }
-
-    return [JSON.parse(resp.body), etag];
-  }
-
-  private async patchJobAndUpdateEtag(
-    id: string,
-    etag: string,
-    patches: JsonPatch[],
-  ): Promise<string> {
-    // Send the patch
-    const job_url = new URL(`api/jobs/${id}`, this.brokerUrl);
-    const resp = await got(job_url, {
-      headers: { etag },
-      json: patches,
-      method: 'PATCH',
-    });
-
-    // Extract the etag header & make sure it was defined
-    const newEtag = resp.headers.etag;
-    if (!newEtag) {
-      throw new Error('missing etag in broker job response');
-    }
-
-    return newEtag;
-  }
-
-  private runBisect(
-    range: string[],
-    gistId: string,
-  ): Promise<FiddleBisectResult> {
-    const [goodVersion, badVersion] = range;
-    // Call fiddle and instruct it to bisect with the supplied parameters
-    return new Promise((resolve, reject) => {
-      execFile(
-        this.fiddleExecPath,
-        ['bisect', goodVersion, badVersion, '--fiddle', gistId] as string[],
-        (err, stdout, stderr) => {
-          // Ensure there was no error
-          if (err === null) {
-            try {
-              // Try to parse the output as well
-              resolve(parseFiddleBisectOutput(stdout));
-            } catch (parseErr) {
-              d('fiddle bisect parse error: %O', parseErr);
-              reject(parseErr);
-            }
-          } else {
-            d(`failed fiddle bisect stdout:\n${stdout}`);
-            d(`failed fiddle bisect stderr:\n${stderr}`);
-            reject(err);
-          }
-        },
-      );
-    });
-  }
-}
-
-// Start the runner and catch any errors that bubble up
-Runner.start().catch((err) => {
+try {
+  const runner = new Runner();
+  runner.start();
+} catch (err) {
   d('encountered an error: %O', err);
-  console.error('execution stopped due to a critical error');
+  console.error('execution stopped due to a critical error', err);
   process.exit(1);
-});
+}
